+++ nerdlets/geo-ops-nerdlet/ViewMap/LocationMetadata.scss
@@ -1,0 +1,29 @@
+ul.detail-panel-metadata-list {
+  margin: 0;
+}
+
+.detail-panel-metadata-item {
+  display: block;
+  border-bottom: 1px dotted #EDEEEE;
+  padding: 8px 16px;
+  list-style-type: none;
+}
+
+.detail-panel-metadata-item-key {
+  width: 150px;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  font-weight: 600;
+  color: #464E4E;
+  text-align: left;
+}
+
+.detail-panel-metadata-item-value {
+  text-align: left;
+  color: #464E4E;
+  white-space: nowrap;
+  width: 100%;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
