+++ nerdlets/shared/components/DetailPanel/DetailPanel.scss
@@ -1,286 +1,3 @@
-.detail-panel-container {
-  height: 100%;
-  background-color: #fff;
-
-  .ic-Icon {
-    line-height: 10px;
-  }
-}
-
-.detail-panel-breadcrumbs-container {
-  display: flex;
-  justify-content: center;
-  padding: 16px;
-  margin-bottom: 0;
-  white-space: nowrap;
-}
-
-.detail-panel-breadcrumbs {
-  margin-bottom: 0;
-}
-
-.detail-panel-breadcrumb {
-  display: inline;
-  position: relative;
-  z-index: 10;
-  padding: 3px 8px;
-  color: #464e4e;
-  margin-right: 3px;
-  font-size: 10px;
-
-  &:first-child {
-    padding-right: 4px;
-    margin-right: 8px;
-    border-radius: 3px 0 0 3px;
-    background-color: #edeeee;
-
-    &:before,
-    &:after {
-      left: 5px;
-    }
-
-    // &:hover {
-    //   transform: translateY(-1px);
-    //   cursor: pointer;
-    //   color: #005054;
-    //   background-color: #D3EFEF;
-
-    //   &:before,
-    //   &:after {
-    //     box-shadow: none;
-    //     background-color: #D3EFEF;
-    //   }
-    //  }
-    
-    &:active {
-      transform: translateY(1px);
-    }
-
-    &.completed-detail-panel-breadcrumb,
-    &.active-step {
-      border-radius: 3px 0 0 3px;
-    }
-  }
-
-  // &:hover {
-  //   color: #005054;
-  //   cursor: pointer;
-  //   transform: translateY(-1px);
-    
-  //   &:before, &:after {
-  //     background-color: #D3EFEF;
-  //   }
-  // }
-  
-  &:active {
-    transform: translateY(1px);
-  }
-
-  span {
-    position: relative;
-    z-index: 10;
-    bottom: 1px;
-    font-size: 14px;
-  }
-}
-
-.detail-panel-breadcrumb:before {
-  content: "";
-  position: absolute;
-  top: 0px;
-  right: 0;
-  left: 0;
-  z-index: -1;
-  width: 100%;
-  height: 50%;
-  background-color: #edeeee;
-  transform: skew(35deg);
-}
-
-.detail-panel-breadcrumb:after {
-  content: "";
-  position: absolute;
-  right: 0;
-  left: 0;
-  bottom: 0.1px;
-  z-index: -1;
-  width: 100%;
-  height: 50%;
-  background-color: #edeeee;
-  transform: skew(-35deg);
-}
-
-.minimized .detail-panel-close-button {
-  display: none;
-}
-
-.detail-panel-minimize-button {
-  display: inline-block;
-  width: 24px;
-  height: 24px;
-  position: relative;
-  margin-left: 4px;
-  top: -1px;
-  border-radius: 3px 0 0 3px;
-  background-color: #edeeee;
-  
-  &:hover {
-    cursor: pointer;
-    background-color: #e3e4e4;
-  }
-  
-  .minimized & {
-    right: 340px;
-  }
-}
-
-.detail-panel-minimize-button-icon {
-  position: relative;
-  top: 3px;
-  left: 9px;
-
-  .minimized & {
-    transform: rotate(180deg) translate(1px, -4px);
-    top: 0;
-  }
-}
-
-.detail-panel-visiblity-controls {
-  position: absolute;
-  right: 0;
-  top: 11px;
-}
-
-a:not([tabindex]).detail-pane-cta {
-  display: inline-block;
-  margin-right: 18px;
-  color: #007e8a;
-  transition: transform .1s ease-out;
-  text-decoration: none;
-
-  &:hover {
-    text-decoration: underline;
-  }
-
-  &:last-child {
-    margin-right: 0px;
-  }
-
-  .ic-Icon {
-    margin-right: 4px;
-    position: relative;
-  }
-}
-
-.detail-panel-header-top-bar-hr {
-  border: none;
-  border-top: 1px dotted rgba(0, 14, 14, 0.1);
-  margin: 0 16px;
-}
-
-.detail-panel-title {
-  margin-bottom: 0;
-  padding: 16px 16px 8px;
-  position: relative;
-
-}
-
-.detail-panel-title-status {
-  display: inline-block;
-  width: 4px;
-  height: 36px;
-  position: absolute;
-  left: 0;
-  top: 12px;
-}
-
-.detail-panel-cta-container {
-  padding: 0 16px 16px;
-
-  &:empty {
-    padding: 0 16px 8px;
-  }
-}
-
-.featured-chart-container {
-  height: 96px;
-  margin-bottom: 16px;
-  padding: 16px;
-}
-
-.detail-panel-featured-chart-header {
-  white-space: nowrap;
-  margin-bottom: 0;
-  font-size: 12px;
-  color: #2a3434;
-}
-
-.detail-panel-featured-chart-query {
-  max-width: 214px;
-  position: relative;
-  bottom: 3px;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-  overflow: hidden;
-  font-size: 10px;
-  font-family: 'fira code', 'open source sans', monospace;
-  color: #007E8A;
-  background-color: #ECF6F6;
-  padding: 4px 4px 3px;
-  border-radius: 4px;
-
-  &:hover {
-    cursor: pointer;
-    color: #005054;
-    background-color: #D3EFEF;
-  }
-}
-
-.detail-panel-featured-chart-comparison-stat {
-  padding: 2px 6px;
-  border-radius: 50px;
-  font-size: 10px;
-
-  &.negative {
-    color: #BF0016;
-    background-color: #FCF2F3;
-  }
-}
-
-.children-container {
-  height: 100%;
-  padding: 16px;
-
-  & > [class*='wnd-Tabs'] {
-    margin: -16px;
-    height: 100%;
-    
-    & > [class*='wnd-Tabs-pane'] {
-      height: 100%;
-    }
-  }
-
-  [class*='Tabs-navigation'] {
-    margin-bottom: 0;
-    background: #FAFBFB;
-    box-shadow: inset 0px 1px 0px #EDEEEE, inset 0px -1px 0px #EDEEEE;
-    border-bottom: 0;
-    padding: 0 8px;
-  }
-
-  [class*='Tab-label'] {
-    font-size: 12px;
-    padding: 9px 0;
-  }
-  
-  [class*='TabContent'] {
-    box-shadow: inset 0 2px 0 #fafbfb;
-    height: calc(100% - 243px);
-    overflow: scroll;
-    padding: 16px;
-  }
-}
-
 // ==============================================================
 // Timeine styles
 // ==============================================================
@@ -626,7 +343,7 @@ a:not([tabindex]).detail-pane-cta {
   transform: rotate(180deg);
 }
 
-.mini-timeline { 
+.mini-timeline {
   &.timeline-container {
     padding: 19px 16px 0;
     box-sizing: border-box;
@@ -710,34 +427,3 @@ a:not([tabindex]).detail-pane-cta {
 }
 
 /*-- end timeline styles --*/
-
-ul.detail-panel-metadata-list {
-  margin: 0;
-}
-
-.detail-panel-metadata-item {
-  display: block;
-  border-bottom: 1px dotted #EDEEEE;
-  padding: 8px 16px;
-  list-style-type: none;
-}
-
-.detail-panel-metadata-item-key {
-  width: 150px;
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-  font-weight: 600;
-  color: #464E4E;
-  text-align: left;
-}
-
-.detail-panel-metadata-item-value {
-  text-align: left;
-  color: #464E4E;
-  white-space: nowrap;
-  width: 100%;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
