+++ nerdlets/shared/components/EntityTypeAhead/EntityTypeAhead.js
@@ -5,39 +5,8 @@ import get from 'lodash.get';
 import { NerdGraphQuery, Spinner } from 'nr1';
 import { NerdGraphError } from '@newrelic/nr1-community';
 
-import styled from 'styled-components';
-
-import { ENTITY_SEARCH_BY_TYPE } from '../services/queries';
-
-const MultiValueRemove = styled.div`
-  width: 22px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border-left: 1px solid #7fcdd5;
-
-  &:hover {
-    cursor: pointer;
-    background-color: rgb(228, 245, 245);
-  }
-
-  [class*='IndicatorsContainer'] & {
-    border: none;
-    transform: scale(1.25);
-    opacity: 0.75;
-    height: 18px;
-    position: relative;
-    right: 9px;
-    width: 18px;
-    background-color: #007e8a1a;
-    border-radius: 100px;
-
-    svg {
-      position: relative;
-      right: 0.5px;
-    }
-  }
-`;
+import { ENTITY_SEARCH_BY_TYPE } from '../../services/queries';
+import { MultiValueRemove } from './styles';
 
 export default class EntityTypeAhead extends React.PureComponent {
   static propTypes = {
