+++ nerdlets/shared/components/MapLocationTable/MapLocationTable.js
@@ -1,31 +1,11 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-import styled from 'styled-components';
 
 import BootstrapTable from 'react-bootstrap-table-next';
 import ToolkitProvider, { Search } from 'react-bootstrap-table2-toolkit';
 import { Icon, Link } from 'nr1';
 
-const StatusColor = styled.div`
-  width: 8px;
-  height: calc(100% + 1px);
-
-  .status-ok & {
-    background-color: #11a600;
-  }
-
-  .status-warning & {
-    background-color: #ffd966;
-  }
-
-  .status-critical & {
-    background-color: #bf0016;
-  }
-
-  .status-not-reporting & {
-    background-color: #8e9494;
-  }
-`;
+import { StatusColor } from './styles';
 
 export default class MapLocationTable extends PureComponent {
   static propTypes = {
