+++ nerdlets/shared/components/Toolbar/ToolbarWrapper.js
@@ -1,29 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Stack, StackItem } from 'nr1';
-import styled from 'styled-components';
-import ToolbarItem from './ToolbarItem';
-
-const StyledStack = styled(Stack)`
-  height: 74px;
-  box-sizing: border-box;
-  border-top: 1px solid #dbdddd;
-  border-bottom: 1px solid #e3e4e4;
-  background: #edeeee;
-  padding: 0 16px;
-`;
-
-const LeftStackItem = styled(StackItem)`
-  ${ToolbarItem}:first-child {
-    padding-left: 0;
-  }
-`;
-
-const RightStackItem = styled(StackItem)`
-  ${ToolbarItem}:last-child {
-    padding-right: 0;
-  }
-`;
+import { Stack } from 'nr1';
+import { StyledStack, LeftStackItem, RightStackItem } from './styles';
 
 const Toolbar = ({ className, left, right }) => (
   <StyledStack
